Move the metal kernels utils in a separate module.
